Split up subcommands into their own directories
