+++ cmd/forklift/cache/cache.go
@@ -1,0 +1,47 @@
+package cache
+
+import (
+	"context"
+	"fmt"
+	"sort"
+
+	units "github.com/docker/go-units"
+	"github.com/pkg/errors"
+	"github.com/urfave/cli/v2"
+
+	"github.com/PlanktoScope/forklift/internal/app/forklift/workspace"
+	"github.com/PlanktoScope/forklift/internal/clients/docker"
+)
+
+// rm
+
+func rmAction(c *cli.Context) error {
+	wpath := c.String("workspace")
+	fmt.Printf("Removing cache from workspace %s...\n", wpath)
+	if err := workspace.RemoveCache(wpath); err != nil {
+		return errors.Wrap(err, "couldn't remove forklift cache of repositories")
+	}
+
+	fmt.Println("Removing unused Docker container images...")
+	client, err := docker.NewClient()
+	if err != nil {
+		return errors.Wrap(err, "couldn't make Docker API client")
+	}
+	report, err := client.PruneUnusedImages(context.Background())
+	if err != nil {
+		return errors.Wrap(err, "couldn't prune unused Docker container images")
+	}
+	sort.Slice(report.ImagesDeleted, func(i, j int) bool {
+		return docker.CompareDeletedImages(report.ImagesDeleted[i], report.ImagesDeleted[j]) < 0
+	})
+	for _, deleted := range report.ImagesDeleted {
+		if deleted.Untagged != "" {
+			fmt.Printf("Untagged %s\n", deleted.Untagged)
+		}
+		if deleted.Deleted != "" {
+			fmt.Printf("Deleted %s\n", deleted.Deleted)
+		}
+	}
+	fmt.Printf("Total reclaimed space: %s\n", units.HumanSize(float64(report.SpaceReclaimed)))
+	return nil
+}
