+++ internal/app/forklift/cli/pkgs.go
@@ -1,0 +1,81 @@
+package cli
+
+import (
+	"fmt"
+	"sort"
+
+	"github.com/PlanktoScope/forklift/internal/app/forklift"
+)
+
+// print
+
+func PrintPkgSpec(indent int, spec forklift.PkgSpec) {
+	IndentedPrintf(indent, "Description: %s\n", spec.Description)
+
+	IndentedPrint(indent, "Maintainers:")
+	if len(spec.Maintainers) == 0 {
+		fmt.Print(" (none)")
+	}
+	fmt.Println()
+	for _, maintainer := range spec.Maintainers {
+		printMaintainer(indent+1, maintainer)
+	}
+
+	if spec.License != "" {
+		IndentedPrintf(indent, "License: %s\n", spec.License)
+	} else {
+		IndentedPrintf(indent, "License: (custom license)\n")
+	}
+
+	IndentedPrint(indent, "Sources:")
+	if len(spec.Sources) == 0 {
+		fmt.Print(" (none)")
+	}
+	fmt.Println()
+	for _, source := range spec.Sources {
+		BulletedPrintf(indent+1, "%s\n", source)
+	}
+}
+
+func printMaintainer(indent int, maintainer forklift.PkgMaintainer) {
+	if maintainer.Email != "" {
+		BulletedPrintf(indent, "%s <%s>\n", maintainer.Name, maintainer.Email)
+	} else {
+		BulletedPrintf(indent, "%s\n", maintainer.Name)
+	}
+}
+
+func PrintDeplSpec(indent int, spec forklift.PkgDeplSpec) {
+	IndentedPrintf(indent, "Deployment:\n")
+	indent++
+
+	// TODO: actually display the definition file?
+	IndentedPrintf(indent, "Definition file: ")
+	if len(spec.DefinitionFile) == 0 {
+		fmt.Println("(none)")
+		return
+	}
+	fmt.Println(spec.DefinitionFile)
+}
+
+func PrintFeatureSpecs(indent int, features map[string]forklift.PkgFeatureSpec) {
+	IndentedPrint(indent, "Optional features:")
+	names := make([]string, 0, len(features))
+	for name := range features {
+		names = append(names, name)
+	}
+	sort.Strings(names)
+	if len(names) == 0 {
+		fmt.Print(" (none)")
+	}
+	fmt.Println()
+	indent++
+
+	for _, name := range names {
+		if description := features[name].Description; description != "" {
+			IndentedPrintf(indent, "%s: %s\n", name, description)
+			continue
+		}
+		IndentedPrintf(indent, "%s\n", name)
+	}
+}
